move: 테스트 service 패키지에서 integration 패키지로 이동
