+++ src/test/java/commaproject/be/commaserver/common/argumentresolver/OauthLoginArgumentResolverTest.java
@@ -2,7 +2,7 @@ package commaproject.be.commaserver.common.argumentresolver;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import commaproject.be.commaserver.service.InitIntegrationTest;
+import commaproject.be.commaserver.integration.InitIntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
