+++ src/test/java/commaproject/be/commaserver/integration/CommentIntegrationTest.java
@@ -1,4 +1,4 @@
-package commaproject.be.commaserver.service;
+package commaproject.be.commaserver.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
